Build temporary game board
This board consists of a status bar that lists the following:
	Round 1 of 5
	Player: 0
	Computer: 0
	Ties: 0

Numbers above are arbitrary.

Placed images of a rock, paper, and a scissor on the page, which will
eventually be clickable options for playing the game.

This will eventually be replaced by JavaScript dynamically creating this
section.  This current game board is a prototype to be used in coding
the JavaScript part.

This still needs the computer choice image and a label/message
instructing the player to choose the rock, paper, or scissors.