Add number validation for createRoundSelector args
numberOfRounds and defaultRound are both checked to be numbers > 1.
If either is invalid, then an error is logged and the bad value is
replaced with 1.

There is also an additional check made for defaultRound, which should be
no greater than numberOfRounds.  If it is, then it is set to be the same
as numberOfRounds.