Add hover and border transition to player options
Hovering over the player's Rock, Paper, or Scissors options now will
give a pointer and change the border color to green.

Also added "no-hover" versions of rock-option, paper-option, and
scissors-option. These will be used after the game ends.

This will also be used to show the computer's selection, which should
not be clickable.