Add hover versions for create RPS functions
Each create Rock/Paper/Scissors set now consists of:

  createScissorsOption
  createScissorsOptionHover
  createScissorsOptionNoHover

The Paper and Scissors functions follow the same pattern.  The
createScissorsOption is the core function that does the main building
part.  The Hover and NoHover add the appropriate class to that.  The
Hover functions will also have Event Listeners added as well.