Add processRound and endGame functions
processRound increments the current round number after each round,
unless it's the final round, then the endGame function is called.

endGame replaces the player choice hover versions of Rock, Paper, and
Scissors with non-hover versions.  No more rounds can be played at this
point and the only option is to click the Reset button to start over.

The displaying of the final results still needs to be implemented.